--- src/pages/setup/index.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import Header from '../../components/setupComponents/Header';
-import Main from '../../components/setupComponents/Main';
-
-const Admin = () => {
-  return (
-    <section>
-        <Header/>
-        <Main/>
-    </section>
-  )
-}
-
-export default Admin
